Separated the OLED related code from keymap.c and moved it to oled_display.c.
