+++ keyboards/helix/rev2/keymaps/default/rules.mk
@@ -26,6 +26,10 @@ LTO_ENABLE = no  # if firmware size over limit, try this option
 #   If OLED_SELECT is other than 'core', use helix/local_drivers/ssd1306.c.
 OLED_SELECT = core
 
+ifeq ($(strip $(OLED_ENABLE)), yes)
+    SRC += oled_display.c
+endif
+
 # convert Helix-specific options (that represent combinations of standard options)
 #   into QMK standard options.
 include $(strip $(KEYBOARD_LOCAL_FEATURES_MK))
